Migrate add tests to docker runner
