+++ e2e-2/add/creates-config-file.test.ts
@@ -1,0 +1,3 @@
+xtest("erros when guet hasn'nt been init", () => {
+    // TODO
+});
