+++ e2e-2/add/errors-on-incorrect-args.test.ts
@@ -1,0 +1,24 @@
+import colors from 'colors';
+
+import { configPath, readJSONFile } from '../../src/utils';
+import { assembleOutput, run } from '../../ts-e2e/utils';
+colors.enable();
+
+test('should fail to add committers when given incorrect args', () => {
+    run('guet init');
+    let [out, statusCode] = run(
+        'guet add fn "full name" fullname@example.com extra'
+    );
+    expect(out).toEqual(assembleOutput(['Too many arguments.'.red]));
+    let found = readJSONFile(configPath);
+    expect(found.committers).toHaveLength(0);
+    expect(statusCode).toEqual(1);
+
+    [out, statusCode] = run(
+        'guet add fn "full name" fullname@example.com extra'
+    );
+    expect(out).toEqual(assembleOutput(['Too many arguments.'.red]));
+    found = readJSONFile(configPath);
+    expect(found.committers).toHaveLength(0);
+    expect(statusCode).toEqual(1);
+});
