--- e2e-2/first.test.ts
@@ -1,3 +1,0 @@
-test('does some thing', () => {
-    expect(0).toEqual(0);
-});
