+++ ts-e2e/utils.ts
@@ -1,8 +1,6 @@
 import { execSync } from 'child_process';
 import fs from 'fs';
 
-import Docker from 'dockerode';
-
 import { configPath, getGitPath } from '../src/utils';
 
 export function run(command: string): [string, number] {
@@ -32,9 +30,4 @@ export function cleanup() {
 
 export function testFileName(name: string) {
     return `test-${name}`;
-}
-
-function dothing() {
-    const h = new Docker();
-    console.log(h);
 }
